dav: properly handle cancellation during mounting
If the user cancels the authentication dialog during the mount
operation, we should catch that case and instead of returing
G_IO_ERROR_FAILED we should G_IO_ERROR_FAILED_HANDLED. This will
prevent an (expected) error being back to the user. This is in
line with what other backends do in the same situation.

https://bugzilla.gnome.org/show_bug.cgi?id=770549